consolidate stuff totally working
